feat: implement TaskRepository persistence layer and test infrastructure
Core Implementation:
- feat(core): add TaskEntity JPA entity with JSON column support
- feat(core): add TaskJpaRepository with custom JPQL queries
- feat(core): add TaskRepositoryImpl implementing domain repository interface
- feat(core): create comprehensive test stub infrastructure

Fixes:
- fix(core): remove @Embeddable from PartInfo for Hibernate 6+ JSON compatibility
- fix(core): change TaskAggregate constructor/setters visibility to public for JPA access
- fix(core): add @Query annotation for TaskJpaRepository.findByFKey method
- fix(core): change ObjectStorageServiceStub adapter type to 'primary'
- fix(core): set Content-Type application/json in FileResponseBuilder.forUpload()
- fix(core): simplify exception creation (remove withoutStack pattern)
- fix(core): change expiry duration types from int to long

Test Infrastructure:
- test(core): create TaskEventPublisherStub for domain events
- test(core): create CallbackTaskPublisherStub for task queue
- test(core): create ObjectStorageServiceStub for in-memory storage
- test(core): create FileSystemStorageStub for filesystem testing
- test(core): exclude Redis auto-configuration in tests
- test(core): add TestStorageConfig with @Primary beans
- test(core): create FileUploadScenarioTest integration test

Technical Details:
- Use @JdbcTypeCode(SqlTypes.JSON) for storing collections/maps as JSON
- TaskEntity maps List<PartInfo>, List<CallbackConfig>, Map context to JSON columns
- Spring Data JPA @Query solves fKey method name parsing issue
- Storage adapter registry uses getAdapterType() as key
- Separate Kafka topics: TaskEventPublisher (events) vs CallbackTaskPublisher (tasks)

+++ file-srv-core/src/main/java/tech/icc/filesrv/config/FileControllerConfig.java
@@ -16,13 +16,13 @@ public class FileControllerConfig {
     private final int maxFileKeyLength;
 
     /** 预签名URL默认有效期（秒） */
-    private final int defaultPresignExpirySeconds;
+    private final long defaultPresignExpirySeconds;
 
     /** 预签名URL最小有效期（秒） */
-    private final int minPresignExpirySeconds;
+    private final long minPresignExpirySeconds;
 
     /** 预签名URL最大有效期（秒） */
-    private final int maxPresignExpirySeconds;
+    private final long maxPresignExpirySeconds;
 
     /** 分页默认大小 */
     private final int defaultPageSize;
